+++ operator/controllers/loki/lokistack_controller_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/go-logr/logr"
 	openshiftconfigv1 "github.com/openshift/api/config/v1"
 	routev1 "github.com/openshift/api/route/v1"
+	cloudcredentialv1 "github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1"
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -202,11 +203,23 @@ func TestLokiStackController_RegisterWatchedResources(t *testing.T) {
 	}
 	table := []test{
 		{
-			src:               &openshiftconfigv1.APIServer{},
+			src:               &cloudcredentialv1.CredentialsRequest{},
 			index:             3,
 			watchesCallsCount: 4,
 			featureGates: configv1.FeatureGates{
 				OpenShift: configv1.OpenShiftFeatureGates{
+					Enabled: true,
+				},
+			},
+			pred: updateOrDeleteOnlyPred,
+		},
+		{
+			src:               &openshiftconfigv1.APIServer{},
+			index:             4,
+			watchesCallsCount: 5,
+			featureGates: configv1.FeatureGates{
+				OpenShift: configv1.OpenShiftFeatureGates{
+					Enabled:          true,
 					ClusterTLSPolicy: true,
 				},
 			},
@@ -214,10 +227,11 @@ func TestLokiStackController_RegisterWatchedResources(t *testing.T) {
 		},
 		{
 			src:               &openshiftconfigv1.Proxy{},
-			index:             3,
-			watchesCallsCount: 4,
+			index:             4,
+			watchesCallsCount: 5,
 			featureGates: configv1.FeatureGates{
 				OpenShift: configv1.OpenShiftFeatureGates{
+					Enabled:      true,
 					ClusterProxy: true,
 				},
 			},