+++ operator/main.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ViaQ/logerr/v2/log"
 	configv1 "github.com/openshift/api/config/v1"
 	routev1 "github.com/openshift/api/route/v1"
+	cloudcredentialv1 "github.com/openshift/cloud-credential-operator/pkg/apis/cloudcredential/v1"
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
@@ -20,6 +21,7 @@ import (
 	lokiv1beta1 "github.com/grafana/loki/operator/apis/loki/v1beta1"
 	lokictrl "github.com/grafana/loki/operator/controllers/loki"
 	"github.com/grafana/loki/operator/internal/config"
+	manifestsocp "github.com/grafana/loki/operator/internal/manifests/openshift"
 	"github.com/grafana/loki/operator/internal/metrics"
 	"github.com/grafana/loki/operator/internal/operator"
 	"github.com/grafana/loki/operator/internal/validation"
@@ -83,6 +85,7 @@ func main() {
 
 		if ctrlCfg.Gates.OpenShift.Enabled {
 			utilruntime.Must(routev1.AddToScheme(scheme))
+			utilruntime.Must(cloudcredentialv1.AddToScheme(scheme))
 		}
 	}
 
@@ -90,6 +93,11 @@ func main() {
 	if err != nil {
 		logger.Error(err, "unable to start manager")
 		os.Exit(1)
+	}
+
+	if ctrlCfg.Gates.OpenShift.Enabled && manifestsocp.DiscoverManagedAuthEnv() != nil {
+		logger.Info("discovered OpenShift Cluster within a managed authentication environment")
+		ctrlCfg.Gates.OpenShift.ManagedAuthEnv = true
 	}
 
 	if err = (&lokictrl.LokiStackReconciler{
@@ -117,6 +125,17 @@ func main() {
 			OperatorNs: ns,
 		}).SetupWithManager(mgr); err != nil {
 			logger.Error(err, "unable to create controller", "controller", "lokistack-dashboards")
+			os.Exit(1)
+		}
+	}
+
+	if ctrlCfg.Gates.OpenShift.ManagedAuthEnabled() {
+		if err = (&lokictrl.CredentialsRequestsReconciler{
+			Client: mgr.GetClient(),
+			Scheme: mgr.GetScheme(),
+			Log:    logger.WithName("controllers").WithName("lokistack-credentialsrequest"),
+		}).SetupWithManager(mgr); err != nil {
+			logger.Error(err, "unable to create controller", "controller", "lokistack-credentialsrequest")
 			os.Exit(1)
 		}
 	}