+++ pkg/storage/bloom/v1/bloom_tokenizer_test.go
@@ -1,7 +1,9 @@
 package v1
 
 import (
+	"context"
 	"fmt"
+	"math"
 	"testing"
 	"time"
 
@@ -9,8 +11,8 @@ import (
 
 	"github.com/grafana/loki/pkg/chunkenc"
 	"github.com/grafana/loki/pkg/logproto"
+	"github.com/grafana/loki/pkg/logql/log"
 	"github.com/grafana/loki/pkg/push"
-	"github.com/grafana/loki/pkg/storage/chunk"
 
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
@@ -33,9 +35,9 @@ var (
 func TestPrefixedKeyCreation(t *testing.T) {
 	var ones uint64 = 0xffffffffffffffff
 
-	ref := logproto.ChunkRef{
-		From:     0,
-		Through:  model.Time(int64(ones)),
+	ref := ChunkRef{
+		Start:    0,
+		End:      model.Time(int64(ones)),
 		Checksum: 0xffffffff,
 	}
 	for _, tc := range []struct {
@@ -86,7 +88,7 @@ func TestSetLineTokenizer(t *testing.T) {
 	require.Equal(t, bt.lineTokenizer.Skip, 7)
 }
 
-func TestPopulateSeriesWithBloom(t *testing.T) {
+func TestTokenizerPopulate(t *testing.T) {
 	var testLine = "this is a log line"
 	bt := NewBloomTokenizer(DefaultNGramLength, DefaultNGramSkip, metrics)
 
@@ -94,23 +96,19 @@ func TestPopulateSeriesWithBloom(t *testing.T) {
 	var lbsList []labels.Labels
 	lbsList = append(lbsList, labels.FromStrings("foo", "bar"))
 
-	var fpList []model.Fingerprint
-	for i := range lbsList {
-		fpList = append(fpList, model.Fingerprint(lbsList[i].Hash()))
-	}
-
-	var memChunks = make([]*chunkenc.MemChunk, 0)
-	memChunk0 := chunkenc.NewMemChunk(chunkenc.ChunkFormatV4, chunkenc.EncSnappy, chunkenc.ChunkHeadFormatFor(chunkenc.ChunkFormatV4), 256000, 1500000)
-	_ = memChunk0.Append(&push.Entry{
+	memChunk := chunkenc.NewMemChunk(chunkenc.ChunkFormatV4, chunkenc.EncSnappy, chunkenc.ChunkHeadFormatFor(chunkenc.ChunkFormatV4), 256000, 1500000)
+	_ = memChunk.Append(&push.Entry{
 		Timestamp: time.Unix(0, 1),
 		Line:      testLine,
 	})
-	memChunks = append(memChunks, memChunk0)
-
-	var chunks = make([]chunk.Chunk, 0)
-	for i := range memChunks {
-		chunks = append(chunks, chunk.NewChunk("user", fpList[i], lbsList[i], chunkenc.NewFacade(memChunks[i], 256000, 1500000), model.TimeFromUnixNano(0), model.TimeFromUnixNano(1)))
-	}
+	itr, err := memChunk.Iterator(
+		context.Background(),
+		time.Unix(0, 0), // TODO: Parameterize/better handle the timestamps?
+		time.Unix(0, math.MaxInt64),
+		logproto.FORWARD,
+		log.NewNoopPipeline().ForStream(nil),
+	)
+	require.Nil(t, err)
 
 	bloom := Bloom{
 		ScalableBloomFilter: *sbf,
@@ -123,12 +121,12 @@ func TestPopulateSeriesWithBloom(t *testing.T) {
 		Series: &series,
 	}
 
-	err := bt.PopulateSeriesWithBloom(&swb, NewSliceIter([][]chunk.Chunk{chunks}))
+	err = bt.Populate(&swb, NewSliceIter([]ChunkRefWithIter{{Ref: ChunkRef{}, Itr: itr}}))
 	require.NoError(t, err)
 	tokenizer := NewNGramTokenizer(DefaultNGramLength, DefaultNGramSkip)
-	itr := tokenizer.Tokens(testLine)
-	for itr.Next() {
-		token := itr.At()
+	toks := tokenizer.Tokens(testLine)
+	for toks.Next() {
+		token := toks.At()
 		require.True(t, swb.Bloom.Test(token))
 	}
 }
@@ -142,23 +140,19 @@ func BenchmarkPopulateSeriesWithBloom(b *testing.B) {
 		var lbsList []labels.Labels
 		lbsList = append(lbsList, labels.FromStrings("foo", "bar"))
 
-		var fpList []model.Fingerprint
-		for i := range lbsList {
-			fpList = append(fpList, model.Fingerprint(lbsList[i].Hash()))
-		}
-
-		var memChunks = make([]*chunkenc.MemChunk, 0)
-		memChunk0 := chunkenc.NewMemChunk(chunkenc.ChunkFormatV4, chunkenc.EncSnappy, chunkenc.ChunkHeadFormatFor(chunkenc.ChunkFormatV4), 256000, 1500000)
-		_ = memChunk0.Append(&push.Entry{
+		memChunk := chunkenc.NewMemChunk(chunkenc.ChunkFormatV4, chunkenc.EncSnappy, chunkenc.ChunkHeadFormatFor(chunkenc.ChunkFormatV4), 256000, 1500000)
+		_ = memChunk.Append(&push.Entry{
 			Timestamp: time.Unix(0, 1),
 			Line:      testLine,
 		})
-		memChunks = append(memChunks, memChunk0)
-
-		var chunks = make([]chunk.Chunk, 0)
-		for i := range memChunks {
-			chunks = append(chunks, chunk.NewChunk("user", fpList[i], lbsList[i], chunkenc.NewFacade(memChunks[i], 256000, 1500000), model.TimeFromUnixNano(0), model.TimeFromUnixNano(1)))
-		}
+		itr, err := memChunk.Iterator(
+			context.Background(),
+			time.Unix(0, 0), // TODO: Parameterize/better handle the timestamps?
+			time.Unix(0, math.MaxInt64),
+			logproto.FORWARD,
+			log.NewNoopPipeline().ForStream(nil),
+		)
+		require.Nil(b, err)
 
 		bloom := Bloom{
 			ScalableBloomFilter: *sbf,
@@ -171,7 +165,7 @@ func BenchmarkPopulateSeriesWithBloom(b *testing.B) {
 			Series: &series,
 		}
 
-		err := bt.PopulateSeriesWithBloom(&swb, NewSliceIter([][]chunk.Chunk{chunks}))
+		err = bt.Populate(&swb, NewSliceIter([]ChunkRefWithIter{{Ref: ChunkRef{}, Itr: itr}}))
 		require.NoError(b, err)
 	}
 }